refactor: #623 — relocate legacy/daemon to top-level src/daemon
DaemonProcess + ServiceManager are independent runtime subsystems used
directly from cli.ts. Same shape as automation/ — top-level alongside
sandbox/, sessions/, etc.

Three cleanups beyond pure relocation:

1. Drop dead legacy/daemon/index.ts barrel — cli.ts imports file paths
   directly, never used the barrel.

2. Drop plain-number PID file fallback in readPidFile. The fallback was
   for migrating users with old-format PID files; pre-users phase, no
   such users exist. Two corresponding test cases removed.

3. Make cliEntry an explicit DaemonOptions field (the previous
   relative-path resolution "path.dirname(import.meta.url) + '..' + 'cli.js'"
   only worked while process.ts lived next to cli.ts in src/legacy/. After
   the move that path resolved to src/cli.js — which doesn't exist. cli.ts
   now resolves its own path and passes cliEntry explicitly. Decouples
   DaemonProcess from filesystem assumptions and unblocks future cli
   relocations.

src/legacy/daemon/ is now gone — legacy/ shrinks to 6 entries.
